+++ CHANGELOG.md
@@ -5,6 +5,7 @@
     - (feature) added new option `cellResizeObserver` for row height auto adjustment
     - fixed minor UI issues for better user experience
     - fixed highlight keywords for async renders
+    - fixed highlight keywords for svg file
 
 + 3.1.1
     - added `defaultPrevented` for event data